+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
+- 2021.01.19, 00:00 -- PD-3538: Webpack: Added sourcemaps for WebUiCore library. WebUiCore-dist.0.2.4. Minor changes. Maintenance tools: `util-update-core-lib.sh` (used in debug mode for build & update WebUiCore library).
+- 2021.01.19, 00:00 -- PD-3538: webpack: Added sourcemaps for WebUiCore library. WebUiCore-dist.0.2.4
 - 2021.01.17, 20:26 -- v.0.0.2: Using WebUiCore dist branch instead separated repository.
 
 <!--
- @changed 2021.01.17, 20:26
+ @changed 2021.01.19, 00:04
 -->